refactor(client): put selector function in a separate file
