+++ packages/client/src/store/selectors.ts
@@ -1,3 +1,4 @@
 import { RootState } from './index'
 
 export const getUser = (state: RootState) => state.user
+export const getPlayersStats = (state: RootState) => state.playersStats
